// src/model/event/event.model.ts
export interface Event {
    ev_id: string;
    ntsb_no: string;
    ev_type: string;
    ev_date: string;
    ev_dow: string;
    ev_time: string;
    ev_tmzn: string;
    ev_city: string;
    ev_state: string;
    ev_country: string;
    ev_site_zipcode: string;
    ev_year: string;
    ev_month: string;
    mid_air: string;
    on_ground_collision: string;
    latitude: string;
    longitude: string;
    latlong_acq: string;
    apt_name: string;
    ev_nr_apt_id: string;
    ev_nr_apt_loc: string;
    apt_dist: string;
    apt_dir: string;
    apt_elev: string;
    wx_brief_comp: string;
    wx_src_iic: string;
    wx_obs_time: string;
    wx_obs_dir: string;
    wx_obs_fac_id: string;
    wx_obs_elev: string;
    wx_obs_dist: string;
    wx_obs_tmzn: string;
    light_cond: string;
    sky_cond_nonceil: string;
    sky_nonceil_ht: string;
    sky_ceil_ht: string;
    sky_cond_ceil: string;
    vis_rvr: string;
    vis_rvv: string;
    vis_sm: string;
    wx_temp: string;
    wx_dew_pt: string;
    wind_dir_deg: string;
    wind_dir_ind: string;
    wind_vel_kts: string;
    wind_vel_ind: string;
    gust_ind: string;
    gust_kts: string;
    altimeter: string;
    wx_dens_alt: string;
    wx_int_precip: string;
    metar: string;
    ev_highest_injury: string;
    inj_f_grnd: string;
    inj_m_grnd: string;
    inj_s_grnd: string;
    inj_tot_f: string;
    inj_tot_m: string;
    inj_tot_n: string;
    inj_tot_s: string;
    inj_tot_t: string;
    invest_agy: string;
    ntsb_docket: string;
    ntsb_notf_from: string;
    ntsb_notf_date: string;
    ntsb_notf_tm: string;
    fiche_number: string;
    lchg_date: string;
    lchg_userid: string;
    wx_cond_basic: string;
    faa_dist_office: string;
}